refactor(components): rename components to use corrext .jsx syntax
