+++ src/components/LoginPage.jsx
@@ -1,13 +1,13 @@
 import { useState } from "react";
 import { useTheme } from 'next-themes';
-import TitleLogo from "@/components/title-logo";
-import LanguageSwitcher from "@/components/language";
+import TitleLogo from "@/components/TitleLogo";
+import LanguageSwitcher from "@/components/LanguageSwitcher";
 import { useTranslation } from "react-i18next";
 import "@/i18n/i18n";
 import { ERROR_CODES } from "@/i18n/errorCodes";
 import ThemeSwitcher from "@/components/Admin/ThemeSwitcher";
 
-export default function Login(props) {
+export default function LoginPage(props) {
   const { t } = useTranslation();
   const { theme } = useTheme();
   const [playerName, setPlayerName] = useState('');
